Moved Action buttons on the "EditExamTemplate" page to the separate component
